Restructure generate command group
