+++ cmd/kro/commands/generate/crd.go
@@ -1,0 +1,73 @@
+// Copyright 2025 The Kube Resource Orchestrator Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package generate
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/kro-run/kro/api/v1alpha1"
+	"github.com/spf13/cobra"
+	"sigs.k8s.io/yaml"
+)
+
+var generateCRDCmd = &cobra.Command{
+	Use:   "crd",
+	Short: "Generate a CustomResourceDefinition (CRD)",
+	Long: "Generate a CustomResourceDefinition (CRD) from a " +
+		"ResourceGraphDefinition file. This command reads the " +
+		"ResourceGraphDefinition and outputs the corresponding CRD " +
+		"in the specified format.",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		if config.resourceGraphDefinitionFile == "" {
+			return fmt.Errorf("ResourceGraphDefinition file is required")
+		}
+
+		data, err := os.ReadFile(config.resourceGraphDefinitionFile)
+		if err != nil {
+			return fmt.Errorf("failed to read ResourceGraphDefinition file: %w", err)
+		}
+
+		var rgd v1alpha1.ResourceGraphDefinition
+		if err = yaml.Unmarshal(data, &rgd); err != nil {
+			return fmt.Errorf("failed to unmarshal ResourceGraphDefinition: %w", err)
+		}
+
+		if err = generateCRD(&rgd); err != nil {
+			return fmt.Errorf("failed to generate CRD: %w", err)
+		}
+
+		return nil
+	},
+}
+
+func generateCRD(rgd *v1alpha1.ResourceGraphDefinition) error {
+	rgdGraph, err := createGraphBuilder(rgd)
+	if err != nil {
+		return fmt.Errorf("failed to setup rgd graph: %w", err)
+	}
+
+	crd := rgdGraph.Instance.GetCRD()
+	crd.SetAnnotations(map[string]string{"kro.run/version": "dev"})
+
+	b, err := marshalObject(crd, config.outputFormat)
+	if err != nil {
+		return fmt.Errorf("failed to marshal CRD: %w", err)
+	}
+
+	fmt.Println(string(b))
+
+	return nil
+}
